+++ ptolemy/math/LongMatrixMath.java
@@ -839,7 +839,7 @@ public class LongMatrixMath {
 
         for (int i = 0; i < rows; i++) {
             for (int j = 0; j < columns; j++) {
-                returnValue[i][j] = new Complex((double) matrix[i][j], 0.0);
+                returnValue[i][j] = new Complex(matrix[i][j], 0.0);
             }
         }
 
@@ -859,7 +859,7 @@ public class LongMatrixMath {
 
         for (int i = 0; i < rows; i++) {
             for (int j = 0; j < columns; j++) {
-                returnValue[i][j] = (double) matrix[i][j];
+                returnValue[i][j] = matrix[i][j];
             }
         }
 
@@ -879,7 +879,7 @@ public class LongMatrixMath {
 
         for (int i = 0; i < rows; i++) {
             for (int j = 0; j < columns; j++) {
-                returnValue[i][j] = (float) matrix[i][j];
+                returnValue[i][j] = matrix[i][j];
             }
         }
