+++ ptolemy/plot/demo/TwoPlotExample.java
@@ -94,25 +94,25 @@ public class TwoPlotExample extends JFrame {
         boolean first = true;
 
         for (int i = 0; i <= 100; i++) {
-            leftPlot.addPoint(0, (double) i, 5 * Math.cos((Math.PI * i) / 20),
+            leftPlot.addPoint(0, i, 5 * Math.cos((Math.PI * i) / 20),
                     !first);
-            leftPlot.addPoint(1, (double) i,
+            leftPlot.addPoint(1, i,
                     4.5 * Math.cos((Math.PI * i) / 25), !first);
-            leftPlot.addPoint(2, (double) i, 4 * Math.cos((Math.PI * i) / 30),
+            leftPlot.addPoint(2, i, 4 * Math.cos((Math.PI * i) / 30),
                     !first);
-            leftPlot.addPoint(3, (double) i,
+            leftPlot.addPoint(3, i,
                     3.5 * Math.cos((Math.PI * i) / 35), !first);
-            leftPlot.addPoint(4, (double) i, 3 * Math.cos((Math.PI * i) / 40),
+            leftPlot.addPoint(4, i, 3 * Math.cos((Math.PI * i) / 40),
                     !first);
-            leftPlot.addPoint(5, (double) i,
+            leftPlot.addPoint(5, i,
                     2.5 * Math.cos((Math.PI * i) / 45), !first);
-            leftPlot.addPoint(6, (double) i, 2 * Math.cos((Math.PI * i) / 50),
+            leftPlot.addPoint(6, i, 2 * Math.cos((Math.PI * i) / 50),
                     !first);
-            leftPlot.addPoint(7, (double) i,
+            leftPlot.addPoint(7, i,
                     1.5 * Math.cos((Math.PI * i) / 55), !first);
-            leftPlot.addPoint(8, (double) i, 1 * Math.cos((Math.PI * i) / 60),
+            leftPlot.addPoint(8, i, 1 * Math.cos((Math.PI * i) / 60),
                     !first);
-            leftPlot.addPoint(9, (double) i,
+            leftPlot.addPoint(9, i,
                     0.5 * Math.cos((Math.PI * i) / 65), !first);
             first = false;
         }
